5.11
Added new dashboard with donut charts and KPI to show the basic information on the top.

Added quick jumps in verification results to prevent long scrolling to the part the user is interested.

Added Donut chart, Bar chart, Radar.

Alot of changes to style.css
Changes to the way .json for reporting is made. Still not final imo, the report is long and unreadable for a user. This is mainly due to the special viz. needing special data. On the other hand the json creation is streamlined and can be reused by any comparator in the future.

#TODO in the mean time I broke show/hide all
Add Correct jcAlgSupport. Tomas Martiska gave a wrong json file. Will probably need a new comparator. Todo 4 more comparators to finish the thesis